requirement.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
boto3==1.41.5
botocore==1.41.5
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-storages==1.14.6
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
msgpack==1.1.2
outcome==1.3.0.post0
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
referencing==0.36.2
rpds-py==0.27.1
s3transfer==0.15.0
selenium==4.38.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.32.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.9.2
types-channels==4.3.0.20250822
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.2.0
zope.interface==8.0.1
